Alumni section Code cleanup/Component creation
