+++ aeroclub/src/components/Pages/Alumni/AlumniData.js
@@ -1,0 +1,201 @@
+import f1 from "../../../images/alumni/sunil saini.jpg";
+import f2 from "../../../images/alumni/Surendra Singh Jayant.jpg";
+import f3 from "../../../images/alumni/SHREY Dixit.jpg";
+import f4 from "../../../images/alumni/Robin Zachariah.jpg";
+import f5 from "../../../images/alumni/rahul thakur.jpg";
+import f6 from "../../../images/alumni/Mohammed Maaz Latifee.jpg";
+import f7 from "../../../images/alumni/Vaibhav Jha.jpg";
+import f8 from "../../../images/alumni/Soumya Khanna.jpeg";
+import f9 from "../../../images/alumni/Aman Sharma(1).jpg";
+import f10 from "../../../images/alumni/sarita chaudhary1.jpeg";
+import f11 from "../../../images/alumni/unnamed.jpg";
+import f12 from "../../../images/alumni/Kishan Tiwari.jfif";
+import f13 from "../../../images/alumni/Aniket Kumar.jpg";
+import f14 from "../../../images/alumni/Bhuvan Jhumb.jpg";
+import f15 from "../../../images/alumni/Apoorva Dinesh Singh.jpg";
+import f16 from "../../../images/alumni/Archit Chaudhary.jpg";
+import f17 from "../../../images/alumni/Ayan Santania.jpg";
+import f18 from "../../../images/alumni/ABHISHEK GAUTAM.jpg";
+
+const members = [
+    {
+      name: "Sunil Kumar saini",
+      yearOfGraduation: 2019,
+      presentOrganisation: "John Deere India Pvt. Ltd.",
+      presentLocation: "Indore, Madhya Pradesh",
+      yourLinkedinId: "https://www.linkedin.com/in/sunilsaini233601",
+      yourContactInformationEmail: "sunilnit22@gmail.com",
+      yourContactInformationPhoneNumber: 8003238261,
+      aPhotoForWebsiteOfYourself: f1,
+    },
+    {
+      name: "Surendra Singh Jayant",
+      yearOfGraduation: 2019,
+      presentOrganisation: "Hindalco",
+      presentLocation: "Hirakud",
+      yourLinkedinId: "https://www.linkedin.com/in/ssjayant",
+      yourContactInformationEmail: "surendramnnit@hotmail.com",
+      yourContactInformationPhoneNumber: 7704874507,
+      aPhotoForWebsiteOfYourself: f2,
+    },
+    {
+      name: "Shrey Dixit",
+      yearOfGraduation: 2020,
+      yourLinkedinId: "linkedin.com/in/shrey-dixit",
+      yourContactInformationEmail: "shreydixit99@gmail.com ",
+      aPhotoForWebsiteOfYourself: f3,
+    },
+    {
+      name: "Robin John Zachariah",
+      yearOfGraduation: 2020,
+      presentOrganisation: undefined,
+      presentLocation: "Kochi, India",
+      yourLinkedinId: "www.linkedin.com/in/robin-john-zachariah",
+      yourContactInformationEmail: "robinjohnzachariah@gmail.com",
+      yourContactInformationPhoneNumber: 7388342800,
+      aPhotoForWebsiteOfYourself: f4,
+    },
+    {
+      name: "Rahul Thakur",
+      yearOfGraduation: 2020,
+      presentOrganisation: "Addverb Technologies",
+      presentLocation: "Noida",
+      yourLinkedinId: "Rahulkrt",
+      yourContactInformationEmail: "Rahulkmrthakur24@gmail.com",
+      aPhotoForWebsiteOfYourself: f5,
+    },
+    {
+      emailAddress: "maaz.latifee2@gmail.com",
+      name: "Mohammed Maaz Latifee",
+      yearOfGraduation: 2020,
+      presentOrganisation: "Infosys",
+      presentLocation: "Abu Dhabi",
+      yourLinkedinId: "linkedin.com/in/mohammed-maaz-latifee-0a3375169",
+      yourContactInformationEmail: "maaz.latifee2@gmail.com",
+      yourContactInformationPhoneNumber: "+91 7753848547",
+      aPhotoForWebsiteOfYourself: f6,
+    },
+    {
+      emailAddress: "vaibhavjha904@gmail.com",
+      name: "Vaibhav Kumar Jha",
+      yearOfGraduation: 2020,
+      presentOrganisation: "Capgemini India",
+      presentLocation: "Kharagpur",
+      yourContactInformationEmail: "vaibhavjha904@gmail.com",
+      yourContactInformationPhoneNumber: 8310731383,
+      yourLinkedinId: "https://www.linkedin.com/in/vaibhav-jha-49745b166",
+      aPhotoForWebsiteOfYourself: f7,
+    },
+    {
+      emailAddress: "soumyakhanna17997@gmail.com",
+      name: "Soumya Khanna",
+      yearOfGraduation: 2020,
+      presentOrganisation: "Bajaj Auto Limited",
+      presentLocation: "Akurdi, Pune",
+      yourLinkedinId: "https://www.linkedin.com/in/soumya-khanna-20169040",
+      yourContactInformationEmail: "soumyakhanna17997@gmail.com",
+      aPhotoForWebsiteOfYourself: f8,
+    },
+    {
+      emailAddress: "sharmaaman0571@gmail.com",
+      name: "Aman sharma",
+      yearOfGraduation: 2020,
+      presentOrganisation: "Hero motocorp",
+      presentLocation: "Gurgaon",
+      yourContactInformationEmail: "sharmaaman0571@gmail.com",
+      yourLinkedinId: "https://www.linkedin.com/in/aman-sharma-910598149",
+      yourContactInformationPhoneNumber: "08218289118",
+      aPhotoForWebsiteOfYourself: f9,
+    },
+    {
+      emailAddress: "saritachaudhary2507@gmail.com",
+      name: "Sarita Chaudhary",
+      yearOfGraduation: 2020,
+      presentOrganisation: "Resonance Eduventures Pvt. Ltd.",
+      presentLocation: "Kota",
+      yourLinkedinId: " https://www.linkedin.com/in/sarita-chaudhary-b92170137/",
+      yourContactInformationEmail: "saritachaudhary2507@gmail.com",
+      yourContactInformationPhoneNumber: "+919554559909",
+      aPhotoForWebsiteOfYourself: f10,
+    },
+    {
+      emailAddress: "bhuvanjhamb29@gmail.com",
+      name: "Ashwini Kumar",
+      yearOfGraduation: 2018,
+      aPhotoForWebsiteOfYourself: f11,
+    },
+    {
+      emailAddress: "Kishan@tsaw.tech",
+      name: "Kishan Tiwari",
+      yearOfGraduation: 2019,
+      presentOrganisation: "TSAW Drones",
+      presentLocation: "Gurgaon",
+      yourLinkedinId: "https://www.linkedin.com/in/kishan-tiwari-poseidon/",
+      yourContactInformationEmail: "kishan@tsaw.tech",
+      yourContactInformationPhoneNumber: 7905461842,
+      aPhotoForWebsiteOfYourself: f12,
+    },
+    {
+      emailAddress: "Kishan@tsaw.tech",
+      name: "Aniket Tiwari",
+      yearOfGraduation: 2019,
+      presentOrganisation: "Assistant Manager, Maruti Suzuki",
+      presentLocation: "Gurgaon",
+      yourLinkedinId: "https://www.linkedin.com/in/aniket-tiwari/",
+      yourContactInformationPhoneNumber: 7905461842,
+      aPhotoForWebsiteOfYourself: f13,
+    },
+    {
+      name: "Bhuvan Jhamb",
+      yearOfGraduation: 2021,
+      presentOrganisation: "Software Developer 2 (Robotics), Intel Labs",
+      presentLocation: "Indore, Madhya Pradesh",
+      yourLinkedinId: "https://www.linkedin.com/in/bhuvanjhamb",
+      yourContactInformationEmail: "bhuvanjhamb29@gmail.com",
+      yourContactInformationPhoneNumber: 999999999,
+      aPhotoForWebsiteOfYourself: f14,
+    },
+    {
+      name: "Apoorva Dinesh Singh",
+      yearOfGraduation: 2021,
+      presentOrganisation: "Software Engineer at Arista Networks",
+      presentLocation: "Indore, Madhya Pradesh",
+      yourLinkedinId: "https://www.linkedin.com/in/apoorva-singh-0111",
+      yourContactInformationEmail: "apoorva011199@gmail.com",
+      yourContactInformationPhoneNumber: 999999999,
+      aPhotoForWebsiteOfYourself: f15,
+    },
+    {
+      name: "ARCHIT CHAUDHARY",
+      yearOfGraduation: 2021,
+      presentOrganisation: undefined,
+      presentLocation: "Indore, Madhya Pradesh",
+      yourLinkedinId: "https://www.linkedin.com/in/archit-chaudhary-010a78164/",
+      yourContactInformationEmail: "chaudharyarchit2000@gmail.com",
+      yourContactInformationPhoneNumber: 999999999,
+      aPhotoForWebsiteOfYourself: f16,
+    },
+    {
+      name: "Ayan Kishore",
+      yearOfGraduation: 2021,
+      presentOrganisation: undefined,
+      presentLocation: "Indore, Madhya Pradesh",
+      yourLinkedinId: "https://www.linkedin.com/",
+      yourContactInformationEmail: "ayan.santania@gmail.com",
+      yourContactInformationPhoneNumber: 999999999,
+      aPhotoForWebsiteOfYourself: f17,
+    },
+    {
+      name: "Abhishek Gautam",
+      yearOfGraduation: 2021,
+      presentOrganisation: "Graduate Engineer Trainee at Hero MotoCorp",
+      presentLocation: "Indore, Madhya Pradesh",
+      yourLinkedinId: "https://www.linkedin.com/in/abhishekgautam01",
+      yourContactInformationEmail: "abhishekgautam689@gmail.com",
+      yourContactInformationPhoneNumber: 999999999,
+      aPhotoForWebsiteOfYourself: f18,
+    }
+  ];
+
+
+  export default members;
